Fleshing out the functionality
All additions are untested.

Added a serial ISR to take an input string, with two numeric values
separated with a 't', where the first corresponds to the desired RPM
of the drive motor, and the second corresponds to the angle in degrees
of the steering servo. E.g '70t15'. This can be entered from the
Arduino serial monitor. This is in place of a ROS command velocity as
implementing that requires the onboard computer etc to be setup, which
i'm yet to do.

Added code to allow the negative RPM of the drive motor, to allow the
robot to move backwards. Can see a lot of errors in this bit.

PID loop needs tuning now.